Modularise the API sub routes into separate files
